source/backend/pyproject.toml
[tool.poetry]
name = "fastapi-fullstack"
version = "0.3.0b"
description = "First backend with using FastAPI framework"
authors = ["monok8i <didkivskiyyura725@gmail.com>"]
readme = "README.md"
packages = [{ include = "fastapi_service" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
pydantic = { extras = ["email"], version = "^2.6.1" }
SQLAlchemy = { extras = ["asyncio"], version = "^2.0.23" }
uvicorn = { extras = ["all"], version = "^0.24.0.post1" }
pydantic-settings = "^2.1.0"
environs = "^9.5.0"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
pyjwt = { extras = ["crypto"], version = "^2.8.0" }
redis = "^5.0.1"
fastapi-cache2 = "^0.2.1"
celery = "^5.3.6"
pika = "^1.3.2"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
bcrypt = "4.0.1"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.1"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '\\.?venv\\.?cert\\.?idea\\.?vscode\\.env\\.?versions\\poetry.lock'


[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "C",
    "DTZ",
    "E",
    "F",
    "I",
    "ICN",
    "ISC",
    "N",
    "PLC",
    "PLE",
    "Q",
    "S",
    "T",
    "W",
    "YTT",
    "RET",
    "SIM",
    "ASYNC",
]
exclude = [
    ".venv",
    ".idea",
    ".tests",
    ".certs",
    "poetry.lock",
]
ignore = []


[tool.mypy]
plugins = [
    "sqlalchemy.ext.mypy.plugin",
    "pydantic.mypy"
]
exclude = [
    ".venv",
    ".idea",
    ".tests",
    ".cert",
]
